+++ src/age_emulator_gb/age_gb_sound_utilities.hpp
@@ -37,8 +37,8 @@ constexpr int8_t gb_sample_cycle_shift = 1; // 2097152 samples per second for ea
 constexpr int32_t gb_cycles_per_sample = 1 << gb_sample_cycle_shift;
 constexpr int32_t gb_cycle_sample_mask = ~(gb_cycles_per_sample - 1);
 
-constexpr uint8 gb_nrX4_length_counter = 0x40;
-constexpr uint8 gb_nrX4_initialize = 0x80;
+constexpr uint8_t gb_nrX4_length_counter = 0x40;
+constexpr uint8_t gb_nrX4_initialize = 0x80;
 
 
 
@@ -57,7 +57,7 @@ public:
         return m_cycles_next_sample;
     }
 
-    void set_channel_multiplier(uint32 channel_multiplier)
+    void set_channel_multiplier(uint32_t channel_multiplier)
     {
         AGE_ASSERT((channel_multiplier & 0xFFFF) <= 8);
         AGE_ASSERT((channel_multiplier >> 16) <= 8);
@@ -65,14 +65,14 @@ public:
         calculate_sample();
     }
 
-    void set_volume(uint8 volume)
+    void set_volume(uint8_t volume)
     {
         AGE_ASSERT(volume <= 60);
         m_volume = volume;
         calculate_sample();
     }
 
-    int32_t generate(pcm_vector &buffer, uint buffer_index, int32_t cycles_elapsed)
+    int32_t generate(pcm_vector &buffer, size_t buffer_index, int32_t cycles_elapsed)
     {
         AGE_ASSERT(m_cycles_per_sample != 0);
         AGE_ASSERT((cycles_elapsed % gb_cycles_per_sample) == 0);
@@ -86,7 +86,7 @@ public:
             int32_t samples_to_write = cycles >> gb_sample_cycle_shift;
             AGE_ASSERT((cycles % gb_cycles_per_sample) == 0);
 
-            for (uint max = buffer_index + samples_to_write; buffer_index < max; ++buffer_index)
+            for (size_t max = buffer_index + samples_to_write; buffer_index < max; ++buffer_index)
             {
                 buffer[buffer_index].m_stereo_sample += m_current_multiplied_sample;
             }
@@ -134,7 +134,7 @@ private:
         //   / 32  ->  4 channels, s0x volume up to 8
         //   / 60  ->  combining volume 1-15 (channels 1,2,4) and volume 0, 1, 0.5, 0.25 (channel 3)
         //
-        int32_t value = (2 * m_current_sample - 15) * std::numeric_limits<int16>::max() * m_volume;
+        int32_t value = (2 * m_current_sample - 15) * std::numeric_limits<int16_t>::max() * m_volume;
         value /= 15 * 32 * 60;
         m_current_multiplied_sample = value * m_channel_multiplier;
     }
@@ -142,11 +142,11 @@ private:
     int32_t m_cycles_per_sample = 0;
     int32_t m_cycles_next_sample = 0;
 
-    uint32 m_channel_multiplier = 0;
-    uint8 m_volume = 15;
+    uint32_t m_channel_multiplier = 0;
+    uint8_t m_volume = 15;
 
-    uint8 m_current_sample = 0;
-    uint32 m_current_multiplied_sample = 0;
+    uint8_t m_current_sample = 0;
+    uint32_t m_current_multiplied_sample = 0;
 };
 
 
@@ -158,15 +158,15 @@ class gb_volume_sweep : public TYPE
 {
 public:
 
-    uint8 read_nrX2() const
+    uint8_t read_nrX2() const
     {
         return m_nrX2;
     }
 
-    bool write_nrX2(uint8 nrX2)
+    bool write_nrX2(uint8_t nrX2)
     {
         // "zombie" update
-        uint volume = m_volume;
+        size_t volume = m_volume;
         if ((m_period == 0) && (m_period_counter > 0))
         {
             ++volume;
@@ -237,7 +237,7 @@ private:
         return result;
     }
 
-    void switch_volume(uint new_volume)
+    void switch_volume(size_t new_volume)
     {
         AGE_ASSERT(new_volume < 0x10);
         if (m_volume != new_volume)
@@ -249,7 +249,7 @@ private:
 
     bool sweep()
     {
-        uint volume = m_sweep_up ? m_volume + 1 : m_volume - 1;
+        size_t volume = m_sweep_up ? m_volume + 1 : m_volume - 1;
 
         bool adjust_volume = volume < 0x10;
         if (adjust_volume)
@@ -260,12 +260,12 @@ private:
         return adjust_volume;
     }
 
-    uint8 m_nrX2 = 0;
+    uint8_t m_nrX2 = 0;
 
     bool m_sweep_up = false;
-    uint m_period = 0;
-    uint m_period_counter = 0;
-    uint m_volume = 0;
+    size_t m_period = 0;
+    size_t m_period_counter = 0;
+    size_t m_volume = 0;
 };
 
 
@@ -277,12 +277,12 @@ class gb_frequency_sweep : public TYPE
 {
 public:
 
-    uint8 read_nrX0() const
+    uint8_t read_nrX0() const
     {
         return m_nrX0;
     }
 
-    bool write_nrX0(uint8 nrX0)
+    bool write_nrX0(uint8_t nrX0)
     {
         m_nrX0 = nrX0 | 0x80;
 
@@ -310,7 +310,7 @@ public:
         bool deactivate = false;
         if (m_shift > 0)
         {
-            uint frequency_bits = sweep_frequency_bits();
+            size_t frequency_bits = sweep_frequency_bits();
             deactivate = invalid_frequency_bits(frequency_bits);
         }
 
@@ -329,7 +329,7 @@ public:
             {
                 if (m_period > 0)
                 {
-                    uint frequency_bits = sweep_frequency_bits();
+                    size_t frequency_bits = sweep_frequency_bits();
                     deactivate = invalid_frequency_bits(frequency_bits);
 
                     if (!deactivate && (m_shift > 0))
@@ -337,7 +337,7 @@ public:
                         m_frequency_bits = frequency_bits;
                         TYPE::set_frequency_bits(m_frequency_bits);
 
-                        uint frequency_bits = sweep_frequency_bits();
+                        size_t frequency_bits = sweep_frequency_bits();
                         deactivate = invalid_frequency_bits(frequency_bits);
                     }
                 }
@@ -350,15 +350,15 @@ public:
 
 private:
 
-    bool invalid_frequency_bits(uint frequency_bits) const
+    bool invalid_frequency_bits(size_t frequency_bits) const
     {
         return frequency_bits > 2047;
     }
 
-    uint sweep_frequency_bits()
+    size_t sweep_frequency_bits()
     {
-        uint shifted = m_frequency_bits >> m_shift;
-        uint result = m_frequency_bits;
+        size_t shifted = m_frequency_bits >> m_shift;
+        size_t result = m_frequency_bits;
 
         if (m_sweep_up)
         {
@@ -378,16 +378,16 @@ private:
         m_period_counter = (m_period == 0) ? 8 : m_period;
     }
 
-    uint8 m_nrX0 = 0;
+    uint8_t m_nrX0 = 0;
 
-    uint m_frequency_bits = 0;
-    uint m_period = 0;
-    uint m_shift = 0;
+    size_t m_frequency_bits = 0;
+    size_t m_period = 0;
+    size_t m_shift = 0;
     bool m_sweep_up = false;
 
     bool m_sweep_enabled = false;
     bool m_swept_down = false;
-    uint m_period_counter = 0;
+    size_t m_period_counter = 0;
 };
 
 
@@ -399,28 +399,28 @@ class gb_wave_generator : public gb_sample_generator<gb_wave_generator>
 public:
 
     gb_wave_generator();
-    gb_wave_generator(uint frequency_counter_shift, uint wave_pattern_index_mask);
+    gb_wave_generator(size_t frequency_counter_shift, size_t wave_pattern_index_mask);
 
-    uint get_frequency_bits() const;
-    uint get_wave_pattern_index() const;
+    size_t get_frequency_bits() const;
+    size_t get_wave_pattern_index() const;
 
-    void set_frequency_bits(uint frequency_bits);
-    void set_low_frequency_bits(uint8 nrX3);
-    void set_high_frequency_bits(uint8 nrX4);
+    void set_frequency_bits(size_t frequency_bits);
+    void set_low_frequency_bits(uint8_t nrX3);
+    void set_high_frequency_bits(uint8_t nrX4);
 
     void reset_wave_pattern_index();
-    void set_wave_pattern_byte(uint offset, uint8 value);
-    void set_wave_pattern_duty(uint8 nrX1);
+    void set_wave_pattern_byte(size_t offset, uint8_t value);
+    void set_wave_pattern_duty(uint8_t nrX1);
 
-    uint8 next_sample();
+    uint8_t next_sample();
 
 private:
 
-    const uint m_frequency_counter_shift;
-    const uint m_index_mask;
+    const size_t m_frequency_counter_shift;
+    const size_t m_index_mask;
 
-    uint m_frequency_bits = 0;
-    uint m_index = 0;
+    size_t m_frequency_bits = 0;
+    size_t m_index = 0;
 
     uint8_array<32> m_wave_pattern;
 };
@@ -433,19 +433,19 @@ class gb_noise_generator : public gb_sample_generator<gb_noise_generator>
 {
 public:
 
-    uint8 read_nrX3() const;
+    uint8_t read_nrX3() const;
 
-    void write_nrX3(uint8 nrX3);
+    void write_nrX3(uint8_t nrX3);
     void init_generator();
 
-    uint8 next_sample();
+    uint8_t next_sample();
 
 private:
 
-    uint8 m_nrX3 = 0;
+    uint8_t m_nrX3 = 0;
     bool m_7steps = false;
     bool m_allow_shift = true;
-    uint16 m_lfsr = 0;
+    uint16_t m_lfsr = 0;
 };
 
 
@@ -456,18 +456,18 @@ class gb_length_counter
 {
 public:
 
-    gb_length_counter(uint counter_mask);
+    gb_length_counter(size_t counter_mask);
 
-    void write_nrX1(uint8 nrX1);
-    bool write_nrX4(uint8 nrX4, bool next_frame_sequencer_step_odd);
+    void write_nrX1(uint8_t nrX1);
+    bool write_nrX4(uint8_t nrX4, bool next_frame_sequencer_step_odd);
     bool cycle();
 
 private:
 
-    const uint m_counter_mask;
+    const size_t m_counter_mask;
 
     bool m_counter_enabled = false;
-    uint m_counter = 0;
+    size_t m_counter = 0;
 };
 
 } // namespace age
